requirements.txt
MLB-StatsAPI>=1.5.0
requests>=2.28.0
websockets>=10.0
cryptography>=38.0.0
python-dotenv>=0.19.0
pandas>=1.5.0
matplotlib>=3.6.0
pytest>=7.0.0
